+++ apk/Changelog.md
@@ -9,9 +9,10 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
-Version September 2023 to February 2024
+Version September 2023 to March 2024
 ==============
 
+* Fix unable to change quality during playback if the default quality in settings was different than AUTO
 * Fix playing single quality streams
 * Fix auto playback not selecting the best possible quality for Live streams, Twitch messed up their bitrate information provided by their servers, with caused issues as the app uses the bitrate and current available internet speed (bandwidth) to determinate what quality to select
 * Fix VOD playback starting from 00:00, after switching apps during a VOD playback the app could lose the VOD time position when you come back to it